Added simple AI to the enemy so they will follow the player around
This is done through an IEnumerator which will hold each of the behaviours and run through their follow methods each game cycle. This also allows me to add more behaviour types later on.

Added collision detection for future use in the Entity Manager class.

Added a method to extensions to let me scale a vector to a defined length.